pyproject.toml
[project]
name = "jwtgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
]
